Added writing sections
I added a large amount of the writing for the project. We still just need to write an overview, descriptions and disccusions (at the bottom) for the infrastructure and police presense sections, and whatever information we want to put in about ethical considerations. There is also an error with the Time of Year plot that should probably be resolved, so I'll commit this to my own branch for now and we can merge it once that's fixed