Added Writing Portions
So I just added the writing portions that I had put on my own branch to master. I'm not sure if there's still errors running some of the code so that should probably be resolved so that the final version has all the images we need. We also still need to add the overview at the beginning and discussion on the last two factors at the bottom.